livekit-plugins-arabic-turn-detector/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "livekit-plugins-arabic-turn-detector"
version = "0.1.0"
description = "Fine-tuned Arabic End-of-Utterance detection for LiveKit voice agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Moustafa3092"},
]
keywords = [
    "livekit",
    "turn-detection",
    "arabic",
    "voice-agent",
    "eou",
    "end-of-utterance",
    "conversational-ai",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "livekit-agents>=1.3.0",
    "livekit-plugins-turn-detector>=1.3.0",
    "transformers>=4.45.0",
    "torch>=2.0.0",
    "huggingface-hub>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/Moustafa3092/livekit-plugins-arabic-turn-detector"
Documentation = "https://github.com/Moustafa3092/livekit-plugins-arabic-turn-detector#readme"
Repository = "https://github.com/Moustafa3092/livekit-plugins-arabic-turn-detector"
"Model on HuggingFace" = "https://huggingface.co/Moustafa3092/livekit-turn-detector-arabic"

[tool.setuptools.packages.find]
where = ["."]
include = ["livekit_plugins_arabic_turn_detector*"]

[tool.setuptools.package-data]
livekit_plugins_arabic_turn_detector = ["py.typed"]
